% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hubs.R
\name{plot_hubs}
\alias{plot_hubs}
\title{Find and plot hub taxa}
\usage{
plot_hubs(graph, bigpoint = 10, littlepoint = 3, layout = "auto")
}
\arguments{
\item{graph}{An igraph object}

\item{bigpoint}{Base point size for hub taxa in the plot}

\item{littlepoint}{Base size for non-hub taxa in the plot}

\item{layout}{Which layout to use for vertices. Currently accepts one of: "auto","gem","davidson"}
}
\value{
plot of igraph network with vertices sized by authority score
}
\description{
Returns a plot with points (vertices) sized by authority score
}
\examples{
plot_hubs(igraph_object)

}
