% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale01.R
\name{scale01}
\alias{scale01}
\title{Scale a numeric vector to between 0 & 1}
\usage{
scale01(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Numeric vector, scaled to 0,1
}
\description{
This can handle vectors full of zeros
}
\examples{
x <- c(0,1,1,2,5,6,7,3)
scale01(x)

}
