% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_keystoneness.R
\name{score_keystoneness}
\alias{score_keystoneness}
\title{Find keystoneness scores for taxa}
\usage{
score_keystoneness(graph, physeq)
}
\arguments{
\item{graph}{An igraph object}

\item{physeq}{A phyloseq object (must include taxonomy table) that matches the taxa in the igraph object}
}
\value{
Data Frame of all taxa identities and keystoneness scores. colNames = c("taxon","keystoneness_score","vertex_id). If the igraph object does not have vertex names, a vector of hub values for taxa over the threshold is returned instead.
}
\description{
Returns a data frame of taxa (vertices) and their keystoneness scores
}
\details{
"keystoneness" is defined here as the sum of zscores for mean degree, inverse betweenness centrality, closeness centrality, and eigenvector centrality. See https://doi.org/10.3389%2Ffmicb.2014.00219 for details.
}
\examples{
score_keystoneness(graph,physeq)

}
