% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_hubs.R
\name{find_hubs}
\alias{find_hubs}
\title{Find hub taxa}
\usage{
find_hubs(graph, physeq, cutoff = "midpoint")
}
\arguments{
\item{graph}{An igraph object}

\item{physeq}{A phyloseq object (must include taxonomy table) that matches the taxa in the igraph object}

\item{cutoff}{Method to use for determining hub taxa authority cutoff. One of "midpoint", "median","1sd","2sd"}
}
\value{
Data Frame of hub taxa identities and hub scores. colNames = c("hub_taxon","hub_score","vertex_id). If the igraph object does not have vertex names, a vector of hub values for taxa over the threshold is returned instead.
}
\description{
Returns a named vector of hub taxa (vertices)
}
\details{
cutoff == "midpoint": hub taxa have scores above the midpoint between min and max hub scores
cutoff == "1sd": hub taxa have scores > 1 std.dev higher than the mean hub score
cutoff == "2sd": hub taxa have scores > 2 std.devs higher than the mean hub score
}
\examples{
find_hubs(igraph_object,physeq)

}
