% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_hubs.R
\name{find_hubs}
\alias{find_hubs}
\title{Find hub taxa}
\usage{
find_hubs(graph, cutoff = "median")
}
\arguments{
\item{graph}{An igraph object}

\item{cutoff}{Method to use for determining hub taxa authority cutoff. One of "midpoint", "median","1sd","2sd}

\item{layout}{Which layout to use for vertices. Currently accepts one of: "auto","gem","davidson"}

\item{littlepoint}{Base size for non-hub taxa in the plot}
}
\value{
plot of igraph network with vertices sized by authority score
}
\description{
Returns a named vector of hub taxa (vertices)
}
\examples{
plot_hubs(igraph_object)

}
